refactor(pages): extract views, drop variant=action, reorder nav tabs
Move watching/events/agents-and-actions implementations into src/views/.
page.tsx files are now one-liner re-exports. Strip variant="action" from
all Button calls throughout. Move Agents & Actions before Events in the
nav registry.
